AI improvement, 특수보상 조건 제대로 적용, 노프렌드 2배로, minor fix
git-svn-id: http://dev.upnl.org/svn/mighty40/trunk@19 9b205077-4d66-49d4-857e-aca6800712ee

+++ SocketBag.cpp
@@ -62,7 +62,7 @@ void CSocketBag::InitForClient( CPlayerSocket* pServerSocket )
 {
 	ASSERT( pServerSocket );
 	m_pServerSocket = pServerSocket;
-	m_pServerSocket->SetTrigger( (DWORD)(LPVOID)this, -1, SockProc );
+	m_pServerSocket->SetTrigger( (DWORD)(LPVOID)this, (DWORD)-1, SockProc );
 }
 
 // 서버용 초기화 (각 플레이어에 대해 반복해서 호출한다)
